Group testcases by view functions
