+++ hc/front/tests/test_add_channel.py
@@ -7,7 +7,7 @@ from hc.test import BaseTestCase
 @override_settings(PUSHOVER_API_TOKEN="token", PUSHOVER_SUBSCRIPTION_URL="url")
 class AddChannelTestCase(BaseTestCase):
 
-    def test_it_works(self):
+    def test_it_adds_email(self):
         url = "/integrations/add/"
         form = {"kind": "email", "value": "alice@example.org"}
 
@@ -51,108 +51,8 @@ class AddChannelTestCase(BaseTestCase):
 
     def test_instructions_work(self):
         self.client.login(username="alice@example.org", password="password")
-        for frag in ("email", "webhook", "pd", "pushover", "hipchat", "victorops"):
+        kinds = ("email", "webhook", "pd", "pushover", "hipchat", "victorops")
+        for frag in kinds:
             url = "/integrations/add_%s/" % frag
             r = self.client.get(url)
             self.assertContains(r, "Integration Settings", status_code=200)
-
-    @override_settings(SLACK_CLIENT_ID=None)
-    def test_slack_webhook_instructions_work(self):
-        self.client.login(username="alice@example.org", password="password")
-        r = self.client.get("/integrations/add_slack/")
-        self.assertContains(r, "Integration Settings", status_code=200)
-
-    @override_settings(SLACK_CLIENT_ID="foo")
-    def test_slack_button(self):
-        self.client.login(username="alice@example.org", password="password")
-        r = self.client.get("/integrations/add_slack/")
-        self.assertContains(r, "slack.com/oauth/authorize", status_code=200)
-
-    @override_settings(SLACK_CLIENT_ID="foo")
-    def test_slack_landing_page(self):
-        r = self.client.get("/integrations/add_slack/")
-        self.assertContains(r, "Before adding Slack integration",
-                            status_code=200)
-
-    def test_it_adds_pushover_channel(self):
-        self.client.login(username="alice@example.org", password="password")
-
-        session = self.client.session
-        session["po_nonce"] = "n"
-        session.save()
-
-        params = "pushover_user_key=a&nonce=n&prio=0"
-        r = self.client.get("/integrations/add_pushover/?%s" % params)
-        assert r.status_code == 302
-
-        channels = list(Channel.objects.all())
-        assert len(channels) == 1
-        assert channels[0].value == "a|0"
-
-    def test_it_validates_pushover_priority(self):
-        self.client.login(username="alice@example.org", password="password")
-
-        session = self.client.session
-        session["po_nonce"] = "n"
-        session.save()
-
-        params = "pushover_user_key=a&nonce=n&prio=abc"
-        r = self.client.get("/integrations/add_pushover/?%s" % params)
-        assert r.status_code == 400
-
-    def test_it_validates_pushover_nonce(self):
-        self.client.login(username="alice@example.org", password="password")
-
-        session = self.client.session
-        session["po_nonce"] = "n"
-        session.save()
-
-        params = "pushover_user_key=a&nonce=INVALID&prio=0"
-        r = self.client.get("/integrations/add_pushover/?%s" % params)
-        assert r.status_code == 403
-
-    def test_it_adds_two_webhook_urls_and_redirects(self):
-        form = {"value_down": "http://foo.com", "value_up": "https://bar.com"}
-
-        self.client.login(username="alice@example.org", password="password")
-        r = self.client.post("/integrations/add_webhook/", form)
-        self.assertRedirects(r, "/integrations/")
-
-        c = Channel.objects.get()
-        self.assertEqual(c.value, "http://foo.com\nhttps://bar.com")
-
-    def test_it_adds_webhook_using_team_access(self):
-        form = {"value_down": "http://foo.com", "value_up": "https://bar.com"}
-
-        # Logging in as bob, not alice. Bob has team access so this
-        # should work.
-        self.client.login(username="bob@example.org", password="password")
-        self.client.post("/integrations/add_webhook/", form)
-
-        c = Channel.objects.get()
-        self.assertEqual(c.user, self.alice)
-        self.assertEqual(c.value, "http://foo.com\nhttps://bar.com")
-
-    def test_it_rejects_non_http_webhook_urls(self):
-        form = {"value_down": "foo", "value_up": "bar"}
-
-        self.client.login(username="alice@example.org", password="password")
-        r = self.client.post("/integrations/add_webhook/", form)
-        self.assertContains(r, "Enter a valid URL.")
-
-        self.assertEqual(Channel.objects.count(), 0)
-
-    def test_it_handles_empty_down_url(self):
-        form = {"value_down": "", "value_up": "http://foo.com"}
-
-        self.client.login(username="alice@example.org", password="password")
-        self.client.post("/integrations/add_webhook/", form)
-
-        c = Channel.objects.get()
-        self.assertEqual(c.value, "\nhttp://foo.com")
-
-    @override_settings(PUSHBULLET_CLIENT_ID="foo")
-    def test_pushbullet_instructions_work(self):
-        self.client.login(username="alice@example.org", password="password")
-        r = self.client.get("/integrations/add_pushbullet/")
-        self.assertContains(r, "www.pushbullet.com/authorize", status_code=200)
